[FIX] update_R bug
This is a fix of two severe bugs that miscalculates the R for some
cells:

This procedure was miscalculated in 2 ways:

1) Due to armadillo's submat the last cell's R of each block was
corrected twice since it was considered as the first cell of next
block.

2) cells_per_block was rounded down resulting in my_ceil not
working. As a result the cells_per_block was one less than it
should. This resulted in up to 1/block_size(by default=0.05), so 19
cells not being corrected each time update_R is invoked. These cells
are random so it does not affect the algorithms ability to perform
corrections it is just adds noise to the gradient descent procedure.